backend/requirements.txt
pycryptodome
pymysql
requests
flask==2.2.5
web3
cryptography
python-dotenv
fastapi
uvicorn
PyJWT